Reverting ShellEscape functionality
The reason for this is that the use case for the functionality was made redundant with
commit 6cb3c226160e4acefe45e75b004ca812eb3e8c53 which changed the dmenu_run script.

git diff -r 66908a748aa8adcc0ac7f1b5faaf62254dbc3c06 -r 6cb3c226160e4acefe45e75b004ca812eb3e8c53

diff --git a/dmenu_run b/dmenu_run
index 22fdd2a..1761c38 100755
--- a/dmenu_run
+++ b/dmenu_run
@@ -1,3 +1,6 @@
 #!/bin/sh
 export _JAVA_AWT_WM_NONREPARENTING=1
-dmenu_path | dmenu -H "${XDG_CACHE_HOME:-$HOME/.cache/}/dmenu_run.hist" "$@" | ${SHELL:-"/bin/sh"} &
\ No newline at end of file
+dmenu_path | dmenu -H "${XDG_CACHE_HOME:-$HOME/.cache/}/dmenu_run.hist" "$@" |
+       while read -r cmd; do
+               echo "$cmd" | xargs setsid -f
+       done
\ No newline at end of file

The problem before was passing plain URLs as command line options to a script which did
not work well when passing to a shell via ${SHELL:-"/bin/sh"}, at least not when passing
to zsh. The shell escape functionality made sure to escape the URL and command before it
was returned and passed to the shell. With the dmenu_run changes above the shell escape
functionality was made redundant hence removing it now.

Can always add it back in again if I find another use case for it.