Significantly improve calcoffsets() performance
getting the width of a string is an O(n) operation, however inside
calcoffsets() we only care about getting the width upto a certain
number.

instead of calling drw_fontset_getwidth() and *then* clamping the
result, this patch introduces drw_fontset_getwidth_clamp() function,
similar to strnlen(), which will stop once we reach n.

this required modification to the drw_text() signature, so a wrapper
function is put in place instead to preserve the api.

when dumping a decently sized (unicode) emoji file into dmenu, I see the
startup time drop significantly with this patch.