mouse support: do not allow interaction with other windows
The previous behaviour was that while dmenu is running the user is
not able to type in other windows because all keypresses goes to
dmenu. But, the user was still allowed to interact with other
windows using the mouse to click links, etc.

This can sometimes lead to confusion where one forgets that dmenu is
running but typing stops working.

This commit includes two practical changes.

1) after grabbing the keyboard we do a one-off attempt at grabbing
the mouse cursor - this prevents e.g. hover effects when moving the
mouse cursor over another window and

2) when clicking outside of the dmenu window dmenu will now close,
but using the scroll wheel or higher buttons will still be ignored

Both of these changes should not apply when dmenu is managed by the
window manager.

If there is a practical need for it both of these changes may be
subject to configuration later down the line.